+++ NEWS.md
@@ -1,3 +1,10 @@
+# mlr3viz 0.2.0
+
+- Added plots for `TaskDens` and `TaskSurv` from package `mlr3proba`.
+- Update documentation of `PredictionRegr` (#23)
+- `autoplot.BenchmarkResult()`: Support for learners with identical IDs (#19)
+- Fixed a bug in `plot_learner_prediction()` (#47)
+
 # mlr3viz 0.1.1
 
 - New plot: learner prediction for objects of class `ResampleResult`.